rename `aux` to `auxiliary`.
`aux` is a reserved in Windows, so using it is not good
for Windows users.
http://en.wikipedia.org/wiki/Filename
