+++ app/in/partake/controller/api/user/GetEventsAPI.java
@@ -11,8 +11,8 @@ import in.partake.model.dao.DAOException;
 import in.partake.model.dao.PartakeConnection;
 import in.partake.model.dao.access.IUserTicketAccess;
 import in.partake.model.dao.access.IEventAccess;
-import in.partake.model.dao.aux.EventFilterCondition;
-import in.partake.model.dao.aux.EventStatus;
+import in.partake.model.dao.auxiliary.EventFilterCondition;
+import in.partake.model.dao.auxiliary.EventStatus;
 import in.partake.model.daofacade.UserDAOFacade;
 import in.partake.model.dto.Event;
 import in.partake.model.dto.EventTicket;
